Seperated the SettingsLeft sections into UserInformation and PasswordReset components
